requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
python-multipart